requirements.txt
ai21==1.2.8
aiodns==3.1.0
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
aleph-alpha-client==3.5.1
anthropic==0.3.5
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.3
blis==0.7.11
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudpathlib==0.15.1
cohere==4.27
confection==0.1.3
cymem==2.0.8
distro==1.8.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
faiss-cpu==1.7.4
fastavro==1.8.2
filelock==3.12.4
Flask==3.0.0
frozenlist==1.4.0
fsspec==2023.9.2
google-ai-generativelanguage==0.3.3
google-api-core==1.34.0
google-auth==2.23.3
google-cloud-aiplatform==1.35.0
google-cloud-bigquery==3.12.0
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.4
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-generativeai==0.2.1
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.3
grpcio-status==1.48.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
networkx==3.1
nltk==3.8.1
numpy==1.26.0
openai==0.28.1
packaging==23.2
pathy==0.10.2
Pillow==10.0.1
preshed==3.0.9
prettytable==3.9.0
proto-plus==1.22.3
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycares==4.4.0
pycparser==2.21
pydantic==1.10.13
pyllms==0.3.3
python-dateutil==2.8.2
python-liquid==1.10.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
shapely==2.0.2
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
thinc==8.2.1
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.34.0
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.0.6
vectordb2==0.1.2
wasabi==1.1.2
wcwidth==0.2.8
weasel==0.3.2
Werkzeug==3.0.0
yarl==1.9.2
zipp==3.17.0
